Cursor/Soysal-Yardimlasma-ve-Bagis-Platformu/backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
cffi==2.0.0
cryptography==46.0.5
Django==6.0.3
django-appconf==1.2.0
django-cors-headers==4.9.0
django-cryptography==1.1
django-filter==25.2
djangorestframework==3.17.0
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
pycparser==3.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
